Enable compiler warnings and fix lots of warnings
Former-commit-id: da76280e99161c35499935c6cc91be628c4b3276

+++ src/datafile.h
@@ -251,7 +251,7 @@ namespace DataFile
             return real();
         }
 
-        void setReal(f64 val)
+        void setReal(f32 val)
         {
             this->~Value();
             dataType = DataType::F32;
